Extract builds and beta app review submission action group actions to separate modules
