+++ src/codemagic/tools/_app_store_connect/action_groups/__init__.py
@@ -1,2 +1,4 @@
 from .app_store_action_group import AppStoreActionGroup
 from .apps_action_group import AppsActionGroup
+from .beta_app_review_submissions_action_group import BetaAppReviewSubmissionsActionGroup
+from .builds_action_group import BuildsActionGroup
